+++ src/components/attribution-editor.js
@@ -15,7 +15,7 @@ class AttributionEditor extends React.Component {
   constructor (props) {
     super(props)
     this.state = {
-      block: this.props.initialBlock
+      block: props.initialBlock
     }
   }
   getChildContext () {
@@ -59,7 +59,6 @@ class AttributionEditor extends React.Component {
     // Send change up to store
     const block = store.routeChange('MEDIA_BLOCK_UPDATE_META', {id, path, value})
     // Send change to view
-    // TODO do this with an ed event / listener in constructor?
     this.setState({block})
   }
   onMoreClick (key) {
@@ -68,12 +67,9 @@ class AttributionEditor extends React.Component {
 
     let block, path, value
 
+    // Send change up to store
     switch (key) {
       case 'delete':
-        // TODO pretty UI
-        if (!confirm('Are you sure?')) {
-          return
-        }
         store.routeChange('MEDIA_BLOCK_REMOVE', id)
         return
       case 'isBasedOnUrl':